# frontend/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "utoipa-axum",
		"description": "Utoipa's axum bindings for seamless integration for the two",
		"contact": {
			"name": "Juha Kukkonen",
			"email": "juha7kukkonen@gmail.com"
		},
		"license": {
			"name": "MIT OR Apache-2.0",
			"identifier": "MIT OR Apache-2.0"
		},
		"version": "0.2.0"
	},
	"paths": {
...
		"/exam/new": {
			"post": {
				"tags": ["Exam"],
				"summary": "Create new exam",
				"operationId": "create",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpsertExamRequestDTO"
							}
						}
					},
					"required": true
				},
				"responses": {
					"201": {
						"description": "Returns created exam's id",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateExamResponseDTO"
								}
							}
						}
					},
					"400": {
						"description": "Wrong data format"
					},
					"401": {
						"description": "No auth data found"
					},
					"403": {
						"description": "User has no permission to create exams"
					}
				},
				"security": [
					{
						"BearerAuth": []
					}
				]
			}
		},
		"/exam/text/new": {
			"post": {
				"tags": ["Exam"],
				"summary": "Create new text",
				"operationId": "create_text",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TextUpsertDTO"
							}
						}
					},
					"required": true
				},
				"responses": {
					"201": {
						"description": "Successfully created text",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TextEntity"
								}
							}
						}
					},
					"400": {
						"description": "Wrong data format"
					},
					"401": {
						"description": "No auth data found"
					},
					"403": {
						"description": "You can't create texts"
					}
				},
				"security": [
					{
						"BearerAuth": []
					}
				]
			}
		},
		"/exam/text/{text_id}": {
			"put": {
				"tags": ["Exam"],
				"summary": "Update text by id",
				"operationId": "update_text",
				"parameters": [
					{
						"name": "text_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TextUpsertDTO"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Successfully updated text",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TextEntity"
								}
							}
						}
					},
					"400": {
						"description": "Wrong data format"
					},
					"401": {
						"description": "No auth data found"
					},
					"403": {
						"description": "You can't update texts"
					},
					"404": {
						"description": "Text not found"
					}
				},
				"security": [
					{
						"BearerAuth": []
					}
				]
			},
			"delete": {
				"tags": ["Exam"],
				"summary": "Delete text by id",
				"operationId": "delete_text",
				"parameters": [
					{
						"name": "text_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"204": {
						"description": "Successfully deleted text"
					},
					"400": {
						"description": "Wrong data format"
					},
					"401": {
						"description": "No auth data found"
					},
					"403": {
						"description": "You can't delete texts"
					},
					"404": {
						"description": "Text not found"
					}
				},
				"security": [
					{
						"BearerAuth": []
					}
				]
			}
		},
		"/exam/{exam_id}": {
			"get": {
				"tags": ["Exam"],
				"summary": "Get exam by id",
				"operationId": "get_by_id",
				"parameters": [
					{
						"name": "exam_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Found exam",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Exam"
								}
							}
						}
					},
					"401": {
						"description": "No auth data found"
					},
					"404": {
						"description": "Exam not found"
					}
				},
				"security": [
					{
						"BearerAuth": []
					}
				]
			},
			"put": {
				"tags": ["Exam"],
				"summary": "Update exam by id",
				"operationId": "update_exam",
				"parameters": [
					{
						"name": "exam_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpsertExamRequestDTO"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Successfully updated exam"
					},
					"400": {
						"description": "Wrong data format"
					},
					"401": {
						"description": "No auth data found"
					},
					"403": {
						"description": "User has no permission to update exam"
					},
					"404": {
						"description": "Exam not found"
					}
				},
				"security": [
					{
						"BearerAuth": []
					}
				]
			},
			"delete": {
				"tags": ["Exam"],
				"summary": "Delete exam by id",
				"operationId": "delete_exam",
				"parameters": [
					{
						"name": "exam_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"204": {
						"description": "Successfully deleted exam"
					},
					"401": {
						"description": "No auth data found"
					},
					"403": {
						"description": "User has no permission to delete exam"
					},
					"404": {
						"description": "Exam not found"
					}
				},
				"security": [
					{
						"BearerAuth": []
					}
				]
			}
		},
...
		"/exam/{exam_id}/entities": {
			"get": {
				"tags": ["Exam"],
				"summary": "Get exam entities (only with active attempt or if exam scores are available or if admin)",
				"operationId": "get_entities",
				"parameters": [
					{
						"name": "exam_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successfully got exam's entities",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/PubExamExtendedEntity"
									}
								}
							}
						}
					},
					"400": {
						"description": "Wrong data format"
					},
					"401": {
						"description": "No auth data found"
					},
					"403": {
						"description": "You have no permission to view entities"
					},
					"404": {
						"description": "Exam not found"
					}
				},
				"security": [
					{
						"BearerAuth": []
					}
				]
			},
			"put": {
				"tags": ["Exam"],
				"summary": "Update exam's entities",
				"operationId": "update_exam_entities",
				"parameters": [
					{
						"name": "exam_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "array",
								"items": {
									"$ref": "#/components/schemas/ExamEntity"
								}
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Successfully updated exam's entities"
					},
					"400": {
						"description": "Wrong data format"
					},
					"401": {
						"description": "No auth data found"
					},
					"403": {
						"description": "User has no permission to update exam"
					},
					"404": {
						"description": "Exam not found"
					}
				},
				"security": [
					{
						"BearerAuth": []
					}
				]
			}
		},
		"/exam/{exam_id}/score": {
			"post": {
				"tags": ["Exam"],
				"summary": "Score all unscored attempts",
				"operationId": "score_unscored",
				"parameters": [
					{
						"name": "exam_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successfully scored attempts in exam"
					},
					"400": {
						"description": "Wrong data format"
					},
					"401": {
						"description": "No auth data found"
					},
					"403": {
						"description": "User has no permission to launch exam scoring"
					},
					"404": {
						"description": "Exam not found"
					}
				},
				"security": [
					{
						"BearerAuth": []
					}
				]
			}
		},
...
		"/lecture/{id}": {
			"get": {
				"tags": ["Lecture"],
				"summary": "Get a lecture by id, including its content and completion state.",
				"operationId": "get_by_id",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Lecture found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LectureResponseDTO"
								}
							}
						}
					},
					"401": {
						"description": "No auth data found"
					},
					"403": {
						"description": "User has no access to this lecture's course"
					},
					"404": {
						"description": "Lecture not found"
					}
				},
				"security": [
					{
						"BearerAuth": []
					}
				]
			},
			"put": {
				"tags": ["Lecture"],
				"summary": "Update a lecture by id.",
				"operationId": "update",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateLectureRequestDTO"
							}
						}
					},
					"required": true
				},
				"responses": {
					"204": {
						"description": "Lecture updated"
					},
					"400": {
						"description": "Invalid request data"
					},
					"401": {
						"description": "No auth data found"
					},
					"403": {
						"description": "User has no permission to update lectures"
					},
					"404": {
						"description": "Lecture or referenced video not found"
					}
				},
				"security": [
					{
						"BearerAuth": []
					}
				]
			},
			"delete": {
				"tags": ["Lecture"],
				"summary": "Delete a lecture by id.",
				"operationId": "delete",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					}
				],
				"responses": {
					"204": {
						"description": "Lecture deleted"
					},
					"401": {
						"description": "No auth data found"
					},
					"403": {
						"description": "User has no permission to delete lectures"
					},
					"404": {
						"description": "Lecture not found"
					}
				},
				"security": [
					{
						"BearerAuth": []
					}
				]
			}
		},
...
		"/report/exam/{exam_id}": {
			"get": {
				"tags": ["Report"],
				"summary": "Exam gradebook: per-attempt scores plus summary statistics.",
				"operationId": "get_exam_gradebook",
				"parameters": [
					{
						"name": "exam_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Exam gradebook",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Gradebook"
								}
							}
						}
					},
					"401": {
						"description": "No auth data found"
					},
					"403": {
						"description": "Only teachers/admins can view reports"
					},
					"404": {
						"description": "Exam not found"
					}
				},
				"security": [
					{
						"BearerAuth": []
					}
				]
			}
		},
...
		"/task/{task_id}": {
			"get": {
				"tags": ["Task"],
				"summary": "Get task by id",
				"operationId": "get_by_id",
				"parameters": [
					{
						"name": "task_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Found task",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PublicTaskDTO"
								}
							}
						}
					},
					"401": {
						"description": "No auth data found"
					},
					"404": {
						"description": "Task not found"
					}
				},
				"security": [
					{
						"BearerAuth": []
					}
				]
			},
			"put": {
				"tags": ["Task"],
				"summary": "Update task by id",
				"operationId": "update_task",
				"parameters": [
					{
						"name": "task_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpsertTaskRequestDTO"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Successfully updated task"
					},
					"400": {
						"description": "Wrong data format"
					},
					"401": {
						"description": "No auth data found"
					},
					"403": {
						"description": "User has no permission to update task"
					},
					"404": {
						"description": "Task not found"
					}
				},
				"security": [
					{
						"BearerAuth": []
					}
				]
			},
			"delete": {
				"tags": ["Task"],
				"summary": "Delete task by id",
				"operationId": "delete_task",
				"parameters": [
					{
						"name": "task_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					}
				],
				"responses": {
					"204": {
						"description": "Successfully deleted task"
					},
					"401": {
						"description": "No auth data found"
					},
					"403": {
						"description": "User has no permission to delete task"
					},
					"404": {
						"description": "Task not found"
					}
				},
				"security": [
					{
						"BearerAuth": []
					}
				]
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"AttemptStatus": {
				"type": "string",
				"enum": ["in_progress", "on_review", "graded"]
			},
...
			"CreateExamResponseDTO": {
				"type": "object",
				"required": ["id"],
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					}
				}
			},
...
			"Exam": {
				"type": "object",
				"required": [
					"id",
					"topic_id",
					"tries_count",
					"duration",
					"name",
					"type",
					"scoring_policy"
				],
				"properties": {
					"description": {
						"type": ["string", "null"]
					},
					"duration": {
						"type": "integer",
						"format": "int32"
					},
					"ends_at": {
						"type": ["string", "null"],
						"format": "date-time"
					},
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"name": {
						"type": "string"
					},
					"scoring_policy": {
						"$ref": "#/components/schemas/ExamScoringPolicy"
					},
					"starts_at": {
						"type": ["string", "null"],
						"format": "date-time"
					},
					"topic_id": {
						"type": "integer",
						"format": "int32"
					},
					"tries_count": {
						"type": "integer",
						"format": "int32"
					},
					"type": {
						"$ref": "#/components/schemas/ExamType"
					}
				}
			},
...
			"ExamEntity": {
				"oneOf": [
					{
						"type": "object",
						"required": ["id", "name"],
						"properties": {
							"id": {
								"type": "integer",
								"format": "int32"
							},
							"name": {
								"type": "string",
								"enum": ["task"]
							}
						}
					},
					{
						"type": "object",
						"required": ["id", "name"],
						"properties": {
							"id": {
								"type": "string",
								"format": "uuid"
							},
							"name": {
								"type": "string",
								"enum": ["text"]
							}
						}
					}
				]
			},
			"ExamScoringPolicy": {
				"type": "string",
				"description": "How a user's multiple attempts at an exam are collapsed into a single score\nfor the rating. Chosen by the teacher when creating/editing the exam.",
				"enum": ["best", "latest", "average"]
			},
			"ExamType": {
				"type": "string",
				"enum": ["Instant", "Delayed"]
			},
...
			"Gradebook": {
				"type": "object",
				"description": "Full exam gradebook: metadata, per-attempt rows and summary stats.",
				"required": [
					"exam_id",
					"exam_name",
					"max_score",
					"tasks",
					"rows",
					"summary"
				],
				"properties": {
					"exam_id": {
						"type": "string",
						"format": "uuid"
					},
					"exam_name": {
						"type": "string"
					},
					"max_score": {
						"type": "integer",
						"format": "int64"
					},
					"rows": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/GradebookRow"
						}
					},
					"summary": {
						"$ref": "#/components/schemas/GradebookSummary"
					},
					"tasks": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/GradebookTask"
						}
					}
				}
			},
			"GradebookRow": {
				"type": "object",
				"description": "One learner attempt in an exam gradebook.",
				"required": [
					"user_id",
					"username",
					"email",
					"attempt_id",
					"started_at",
					"ends_at",
					"score",
					"status",
					"task_scores"
				],
				"properties": {
					"attempt_id": {
						"type": "string",
						"format": "uuid"
					},
					"email": {
						"type": "string"
					},
					"ends_at": {
						"type": "string",
						"format": "date-time"
					},
					"first_name": {
						"type": ["string", "null"]
					},
					"last_name": {
						"type": ["string", "null"]
					},
					"patronymic": {
						"type": ["string", "null"]
					},
					"score": {
						"type": "number",
						"format": "double"
					},
					"started_at": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"$ref": "#/components/schemas/AttemptStatus"
					},
					"task_scores": {
						"type": "object",
						"description": "Per-task score for this attempt, keyed by task id (as string).",
						"additionalProperties": {
							"type": "number",
							"format": "double"
						},
						"propertyNames": {
							"type": "string"
						}
					},
					"user_id": {
						"type": "string",
						"format": "uuid"
					},
					"username": {
						"type": "string"
					}
				}
			},
			"GradebookSummary": {
				"type": "object",
				"description": "Aggregate statistics across all attempts of an exam.",
				"required": [
					"total_attempts",
					"participants",
					"graded",
					"on_review",
					"in_progress",
					"average_score",
					"highest_score",
					"lowest_score"
				],
				"properties": {
					"average_score": {
						"type": "number",
						"format": "double"
					},
					"graded": {
						"type": "integer",
						"minimum": 0
					},
					"highest_score": {
						"type": "number",
						"format": "double"
					},
					"in_progress": {
						"type": "integer",
						"minimum": 0
					},
					"lowest_score": {
						"type": "number",
						"format": "double"
					},
					"on_review": {
						"type": "integer",
						"minimum": 0
					},
					"participants": {
						"type": "integer",
						"minimum": 0
					},
					"total_attempts": {
						"type": "integer",
						"minimum": 0
					}
				}
			},
			"GradebookTask": {
				"type": "object",
				"description": "A task in an exam, used for the per-task export columns.",
				"required": ["id", "title", "max_score"],
				"properties": {
					"id": {
						"type": "integer",
						"format": "int64"
					},
					"max_score": {
						"type": "integer",
						"format": "int64"
					},
					"title": {
						"type": "string"
					}
				}
			},
...
			"LectureResponseDTO": {
				"type": "object",
				"description": "Full lecture view, including the rich-text body and completion state.\n\nThe `video_id`, when present, is resolved to a playable URL through the\nexisting `GET /video/{video_id}` endpoint.",
				"required": ["id", "title", "completed"],
				"properties": {
					"completed": {
						"type": "boolean"
					},
					"content": {
						"type": ["string", "null"]
					},
					"description": {
						"type": ["string", "null"]
					},
					"id": {
						"type": "integer",
						"format": "int32"
					},
					"title": {
						"type": "string"
					},
					"video_id": {
						"type": ["string", "null"]
					}
				}
			},
...
			"PubExamExtendedEntity": {
				"oneOf": [
					{
						"type": "object",
						"required": ["task", "type"],
						"properties": {
							"task": {
								"$ref": "#/components/schemas/PublicTaskDTO"
							},
							"type": {
								"type": "string",
								"enum": ["task"]
							}
						}
					},
					{
						"type": "object",
						"required": ["text", "type"],
						"properties": {
							"text": {
								"$ref": "#/components/schemas/TextEntity"
							},
							"type": {
								"type": "string",
								"enum": ["text"]
							}
						}
					}
				]
			},
...
			"PublicTaskDTO": {
				"type": "object",
				"required": ["id", "title", "task_type", "points", "configuration"],
				"properties": {
					"configuration": {
						"$ref": "#/components/schemas/PublicTaskConfig"
					},
					"description": {
						"type": ["string", "null"]
					},
					"id": {
						"type": "integer",
						"format": "int64"
					},
					"points": {
						"type": "integer",
						"format": "int64"
					},
					"task_type": {
						"$ref": "#/components/schemas/TaskType"
					},
					"title": {
						"type": "string"
					}
				}
			},
...
			"TaskType": {
				"type": "string",
				"enum": [
					"SingleChoice",
					"MultipleChoice",
					"ShortText",
					"LongText",
					"Ordering",
					"FileUpload",
					"ctfd"
				]
			},
...
			"TextEntity": {
				"type": "object",
				"required": ["id", "text"],
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"text": {
						"type": "string"
					}
				}
			},
			"TextUpsertDTO": {
				"type": "object",
				"required": ["text"],
				"properties": {
					"text": {
						"type": "string"
					}
				}
			},
...
			"UpdateLectureRequestDTO": {
				"type": "object",
				"required": ["title", "order_index"],
				"properties": {
					"content": {
						"type": ["string", "null"]
					},
					"description": {
						"type": ["string", "null"]
					},
					"order_index": {
						"type": "integer",
						"format": "int32"
					},
					"title": {
						"type": "string"
					},
					"video_id": {
						"type": ["string", "null"]
					}
				}
			},
...
			"UpsertExamRequestDTO": {
				"type": "object",
				"required": ["topic_id", "name", "tries_count", "duration", "type"],
				"properties": {
					"description": {
						"type": ["string", "null"]
					},
					"duration": {
						"type": "integer",
						"format": "int32"
					},
					"ends_at": {
						"type": ["string", "null"],
						"format": "date-time"
					},
					"name": {
						"type": "string"
					},
					"scoring_policy": {
						"$ref": "#/components/schemas/ExamScoringPolicy"
					},
					"starts_at": {
						"type": ["string", "null"],
						"format": "date-time"
					},
					"topic_id": {
						"type": "integer",
						"format": "int32"
					},
					"tries_count": {
						"type": "integer",
						"format": "int32"
					},
					"type": {
						"$ref": "#/components/schemas/ExamType"
					}
				}
			},
			"UpsertTaskRequestDTO": {
				"type": "object",
				"required": ["title", "task_type", "points", "configuration"],
				"properties": {
					"configuration": {
						"$ref": "#/components/schemas/TaskConfig"
					},
					"description": {
						"type": ["string", "null"]
					},
					"points": {
						"type": "integer",
						"format": "int32"
					},
					"task_type": {
						"$ref": "#/components/schemas/TaskType"
					},
					"title": {
						"type": "string"
					}
				}
			},
...
		}
	}
}